refactor(node): further modularize logging
