+++ sn_node/src/bin/sn_node/main.rs
@@ -27,8 +27,6 @@
     unused_results
 )]
 
-#[cfg(not(feature = "tokio-console"))]
-use sn_interface::LogFormatter;
 use sn_node::node::{Config, Error as NodeError, Event, MembershipEvent, NodeApi};
 
 use clap::{CommandFactory, Parser};
@@ -37,24 +35,11 @@ use color_eyre::{Section, SectionExt};
 #[cfg(not(feature = "tokio-console"))]
 use eyre::Error;
 use eyre::{eyre, Context, ErrReport, Result};
-use file_rotate::{
-    compression::Compression,
-    suffix::{AppendTimestamp, FileLimit},
-    ContentLimit,
-};
 use self_update::{cargo_crate_version, Status};
 use std::{io::Write, process::exit};
 use tokio::time::{sleep, Duration};
 use tracing::{self, debug, error, info, trace, warn};
 
-#[cfg(not(feature = "tokio-console"))]
-use tracing_appender::non_blocking::WorkerGuard;
-
-#[cfg(not(feature = "tokio-console"))]
-use tracing_subscriber::filter::EnvFilter;
-
-#[cfg(not(feature = "tokio-console"))]
-const MODULE_NAME: &str = "sn_node";
 const JOIN_TIMEOUT_SEC: u64 = 30;
 const BOOTSTRAP_RETRY_TIME_SEC: u64 = 5;
 
@@ -75,7 +60,7 @@ fn main() -> Result<()> {
     trace!("Initial node config: {config:?}");
 
     #[cfg(not(feature = "tokio-console"))]
-    let _guard = init_node_logging(config).map_err(Error::from)?;
+    let _guard = log::init_node_logging(config).map_err(Error::from)?;
 
     loop {
         create_runtime_and_node()?;
@@ -126,96 +111,6 @@ fn create_runtime_and_node() -> Result<()> {
     debug!("Node runtime should be shutdown now");
     Ok(())
 }
-
-/// Inits node logging, returning the global node guard if required.
-/// This guard should be held for the life of the program.
-///
-/// Logging should be instantiated only once.
-#[cfg(not(feature = "tokio-console"))]
-fn init_node_logging(config: Config) -> Result<Option<WorkerGuard>> {
-    // ==============
-    // Set up logging
-    // ==============
-
-    let mut _optional_guard: Option<WorkerGuard> = None;
-
-    let filter = match EnvFilter::try_from_env("RUST_LOG") {
-        Ok(filter) => filter,
-        // If we have an error (ie RUST_LOG not set or otherwise), we check the verbosity flags
-        Err(_) => {
-            // we manually determine level filter instead of using tracing EnvFilter.
-            let level_filter = config.verbose();
-            let module_filter = format!("{}={}", MODULE_NAME, level_filter)
-                .parse()
-                .wrap_err("BUG: invalid module filter constructed")?;
-            EnvFilter::from_default_env().add_directive(module_filter)
-        }
-    };
-
-    _optional_guard = if let Some(log_dir) = config.log_dir() {
-        println!("Starting logging to directory: {:?}", log_dir);
-
-        let mut content_limit = ContentLimit::BytesSurpassed(config.logs_max_bytes);
-        if config.logs_max_lines > 0 {
-            content_limit = ContentLimit::Lines(config.logs_max_lines);
-        }
-
-        // FileRotate crate changed `0 means for all` to `0 means only original`
-        // Here set the retained value to be same as uncompressed in case of 0.
-        let logs_retained = if config.logs_retained == 0 {
-            config.logs_uncompressed
-        } else {
-            config.logs_retained
-        };
-        let file_appender = log::FileRotateAppender::make_rotate_appender(
-            log_dir,
-            "sn_node.log",
-            AppendTimestamp::default(FileLimit::MaxFiles(logs_retained)),
-            content_limit,
-            Compression::OnRotate(config.logs_uncompressed),
-        );
-
-        // configure how tracing non-blocking works: https://tracing.rs/tracing_appender/non_blocking/struct.nonblockingbuilder#method.default
-        let non_blocking_builder = tracing_appender::non_blocking::NonBlockingBuilder::default();
-
-        let (non_blocking, guard) = non_blocking_builder
-            // lose lines and keep perf, or exert backpressure?
-            .lossy(false)
-            // optionally change buffered lines limit
-            // .buffered_lines_limit(buffered_lines_limit)
-            .finish(file_appender);
-
-        let builder = tracing_subscriber::fmt()
-            // eg : RUST_LOG=my_crate=info,my_crate::my_mod=debug,[my_span]=trace
-            .with_env_filter(filter)
-            .with_thread_names(true)
-            .with_ansi(false)
-            .with_writer(non_blocking);
-
-        if config.json_logs {
-            builder.json().init();
-        } else {
-            builder.event_format(LogFormatter::default()).init();
-        }
-
-        Some(guard)
-    } else {
-        println!("Starting logging to stdout");
-
-        tracing_subscriber::fmt()
-            .with_thread_names(true)
-            .with_ansi(false)
-            .with_env_filter(filter)
-            .with_target(false)
-            .event_format(LogFormatter::default())
-            .init();
-
-        None
-    };
-
-    Ok(_optional_guard)
-}
-
 
 
 async fn run_node(config: Config) -> Result<()> {
